sam/thumbnail/requirements.txt
boto3
pillow>=8.0
requests
